% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_timewindow.R
\name{impute_timewindow}
\alias{impute_timewindow}
\title{Impute Missing Values Within a Time Window}
\usage{
impute_timewindow(
  data,
  variable,
  patient_id_col,
  dy_col,
  window_days = 30,
  is_date = FALSE
)
}
\arguments{
\item{data}{A data frame containing repeated measures data with a time variable.}

\item{variable}{A character string specifying the variable to impute (e.g., `"hba1c"`).}

\item{patient_id_col}{A character string indicating the name of the patient ID column (e.g., `"id"`).}

\item{dy_col}{A character string indicating the name of the time column (e.g., `"dy"` for days from baseline).}

\item{window_days}{Numeric. The maximum absolute time difference (in days) allowed for imputation. Default is 30. Using a window_days value greater than 30 is not recommended for imputation of biological variables that can change overtime such as HbA1c, insulin dose or anthropometric values.}

\item{is_date}{Logical. If TRUE, the `dy_col` is a `Date` object and will be converted to numeric days. If FALSE, assumes `dy_col` is already numeric. Default is FALSE.}
}
\value{
A cleaned and augmented data frame containing:
- The original variable
- A new column `imputed_value_variable` containing the imputed values
- A logical flag `variable_carried` indicating whether imputation was applied
}
\description{
This function imputes missing values in a longitudinal dataset by carrying forward or backward
the nearest observed non-missing value within a defined number of days (default is 30). It is
suitable for clinical measures such as HbA1c, insulin, or any time-varying biomarker. It also
can be used to impute antropometric data in adults over the age of 20.

The function searches within each patient for the closest available value in time and imputes it
only if it falls within the defined time window. This is useful for preserving temporal integrity
while minimizing missingness in sparse clinical data.
}
\details{
- Only missing values with a non-missing neighbor within the specified window are imputed.
- The nearest value (forward or backward in time) is used.
- Imputation does not occur if no observed value is found within the window.
- The function is flexible and can be used for both lab and clinical measurements.
- Ensure you assign the function in order to store the outputted dataframe in your environment.
}
\examples{
example_data <- data.frame(
  id = c(1, 1, 1, 2, 2),
  dy = c(0, 15, 45, 0, 20),
  hba1c = c(7.2, NA, 7.4, NA, 6.9)
)
impute_timewindow(
  example_data,
  variable = "hba1c",
  patient_id_col = "id",
  dy_col = "dy",
  is_date = FALSE
)
}
\author{
Alice Carr
}
